+++ java/sui-entity/src/main/resources/db/migration_sui/V3__add__color_setting__to__table_info.sql
@@ -1,0 +1,2 @@
+ALTER TABLE sui_meta.table_info DROP COLUMN IF EXISTS color_settings;
+ALTER TABLE sui_meta.table_info ADD color_settings JSON;