+++ packages/SUI.BASE_COMPONENTS/src/Dnf/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./MetaSettingTableRowColorRowComponent";
+export * from "./ParameterRowComponent";