+++ packages/SUI.BASE_COMPONENTS/src/Dnf/index.ts
@@ -3,4 +3,5 @@ export * from "./AndForm";
 export * from "./AndFormRowElement";
 export * from "./DnfForm";
 export * from "./ICommonDnfProps";
+export * from "./components";
 export * from "./utils";