+++ packages/SUI.BASE_COMPONENTS/src/Dnf/utils/DnfUtils.tsx
@@ -1,7 +1,10 @@
-import {DatePicker, Icon, Input, InputNumber, Select, Switch} from 'antd';
+import {DatePicker, Form, Icon, Input, InputNumber, Select, Switch} from 'antd';
+import {FormItemProps} from "antd/lib/form";
+import {GetFieldDecoratorOptions, WrappedFormUtils} from "antd/lib/form/Form";
+import {OptionProps, SelectProps} from "antd/lib/select";
 import * as React from 'react';
 
-import {ActionType, FilterType, isListAction, isMomentType} from '../../utils';
+import {ActionType, FilterType, getRussianName, isListAction, isMomentType} from '../../utils';
 
 export function getActions(filterType?: FilterType, constant?: boolean): ActionType[] {
   let actions: ActionType[] = [ActionType.EQUAL];
@@ -42,7 +45,7 @@ export function getActions(filterType?: FilterType, constant?: boolean): ActionT
   return actions;
 }
 
-export function getInputElement<Props>(type: FilterType, action: ActionType, props: Props): JSX.Element {
+export function getInputElement<Props>(type: FilterType | undefined, action: ActionType | undefined, props: Props): JSX.Element {
   if (isListAction(action)) {
     return (
       <Select
@@ -78,11 +81,74 @@ export function getInputElement<Props>(type: FilterType, action: ActionType, pro
   }
 }
 
-export function getValuePropName(type: FilterType): string {
+export function getValuePropName(type: FilterType | undefined): string {
   switch (type) {
     case FilterType.BOOLEAN:
       return "checked";
     default:
       return "value";
   }
+}
+
+export function generateSelectFormItem(
+  form: WrappedFormUtils,
+  decoratorName: string,
+  selectData: OptionProps[],
+  props?: {
+    decoratorOptions?: GetFieldDecoratorOptions;
+    formItemProps?: FormItemProps;
+    selectProps?: SelectProps<ActionType>;
+  }): JSX.Element {
+  const {getFieldDecorator} = form;
+  const safeProps = props || {};
+
+  return (
+    <Form.Item
+      style={{marginBottom: 0}}
+      {...safeProps.formItemProps}
+    >
+      {getFieldDecorator(decoratorName, {
+        ...safeProps.decoratorOptions
+      })(
+        <Select
+          style={{width: "100%"}}
+          {...safeProps.selectProps}
+        >
+          {selectData.map(optionProps => (
+            <Select.Option
+              {...optionProps}
+            />
+          ))}
+        </Select>
+      )}
+    </Form.Item>
+  );
+}
+
+export function generateActionSelectFormItem(
+  form: WrappedFormUtils,
+  decoratorName: string,
+  actions: ActionType[],
+  props?: {
+    decoratorOptions?: GetFieldDecoratorOptions;
+    formItemProps?: FormItemProps;
+    selectProps?: SelectProps<ActionType>;
+  }): JSX.Element {
+  const safeProps = props || {};
+
+  return generateSelectFormItem(
+    form,
+    decoratorName,
+    actions.map(actionType => ({value: actionType, children: getRussianName(actionType)})),
+    {
+      ...safeProps,
+      decoratorOptions: {
+        initialValue: ActionType.EQUAL,
+        normalize: (value): ActionType => value && actions.includes(value)
+          ? value
+          : ActionType.EQUAL,
+        ...safeProps.decoratorOptions
+      }
+    }
+  );
 }