+++ packages/SUI.BASE_COMPONENTS/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./actionType";
+export * from "./condition";
 export * from "./draw";
 export * from "./filterType";