+++ packages/SUI.CHARTS/src/ReportTop3DBarChart.tsx
@@ -17,9 +17,10 @@ interface ITop3DBarChartProps {
   data?: any[];
   decimalValues?: boolean;
   labelPanelWidth?: number;
+  labelTemplate?: string;
   maxValue?: number;
   // Relative - 0 to 100
-  type: "relative" | "absolute";
+  type: "relative" | "absolute" | string;
   valueDataField: string;
 
   // tslint:disable-next-line:no-any
@@ -52,16 +53,18 @@ export class ReportTop3DBarChart extends React.Component<ITop3DBarChartProps, {
 
     return (
       <div style={{display: 'flex'}}>
-        <div style={{width: this.props.labelPanelWidth || defaultLabelPanelWidth, marginTop: 30, height: 270}}>
-          {data.map((element: IObjectWithIndex, index) => (<div
-            key={index.toString()}
-            style={{height: 216 / (length || 1), display: 'flex', alignItems: 'center', ...labelStyle}}
-          >
-            <div style={{width: '100%'}}>
-              {this.props.categoryAxisLabelGenerator(element)}
-            </div>
-          </div>))}
-        </div>
+        {!!length && (
+          <div style={{width: this.props.labelPanelWidth || defaultLabelPanelWidth, marginTop: 30, height: 270}}>
+            {data.map((element: IObjectWithIndex, index) => (<div
+              key={index.toString()}
+              style={{height: 216 / (length || 1), display: 'flex', alignItems: 'center', ...labelStyle}}
+            >
+              <div style={{width: '100%'}}>
+                {this.props.categoryAxisLabelGenerator(element)}
+              </div>
+            </div>))}
+          </div>
+        )}
         {this.props.data && (
           <XYChart3DWrapper
             style={{flexGrow: 1, height: 300}}
@@ -78,6 +81,7 @@ export class ReportTop3DBarChart extends React.Component<ITop3DBarChartProps, {
               const valueAxis = chart.xAxes.push(new am4charts.ValueAxis());
               valueAxis.renderer.minGridDistance = 40;
               valueAxis.min = 0;
+              valueAxis.maxPrecision = 0;
 
               if (this.props.type === "relative") {
                 valueAxis.max = 99.99999999;
@@ -95,6 +99,7 @@ export class ReportTop3DBarChart extends React.Component<ITop3DBarChartProps, {
 
               const valueLabel = series.bullets.push(new am4charts.LabelBullet());
               valueLabel.label.text = `{valueX${this.props.decimalValues ? '.formatNumber("#.##")' : ''}}`;
+              valueLabel.label.text = this.props.labelTemplate || `{valueX${this.props.decimalValues ? '.formatNumber("#.##")' : ''}}`;
               valueLabel.label.fontSize = 16;
               valueLabel.label.truncate = false;
               valueLabel.label.hideOversized = false;
@@ -106,11 +111,11 @@ export class ReportTop3DBarChart extends React.Component<ITop3DBarChartProps, {
       </div>);
   }
 
-  private mapData(data: IObjectWithIndex[]): IObjectWithIndex[] {
+  private mapData(data: IObjectWithIndex[], maxValue: number): IObjectWithIndex[] {
     return data.map((element) => {
       let value = element[this.props.valueDataField];
       if (this.props.type === "absolute") {
-        value = value / (this.state && this.state.maxValue || 1) * 100;
+        value = value / (maxValue || 1) * 100;
       }
 
       return {
@@ -124,13 +129,14 @@ export class ReportTop3DBarChart extends React.Component<ITop3DBarChartProps, {
 
   private updateState(): void {
     const data = this.props.data || [];
+    const valueDataField = this.props.valueDataField;
+
     let maxValue: number | undefined;
     if (this.props.type === "absolute") {
-      const valueDataField = this.props.valueDataField;
-      maxValue = data
-        .reduce((previousValue, currentValue) => (previousValue[valueDataField] || 0) < currentValue[valueDataField] ? currentValue : previousValue, {});
+      maxValue = Math.max(...data.map(element => element[valueDataField]), 0);
     }
-    this.setState({mappedData: this.mapData(data).reverse(), maxValue});
+
+    this.setState({mappedData: this.mapData(data, maxValue as number).reverse(), maxValue});
   }
 
 }