+++ packages/SUI.CHARTS/CHANGELOG.md
@@ -3,9 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [5.9.8](https://github.com/mbuyakov/SUI.CORE/compare/v5.9.7...v5.9.8) (2019-09-23)
+# [5.12.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.11.0...v5.12.0) (2019-09-24)
+
+
+### Bug Fixes
+
+* chart absolute color fix ([62f2355](https://github.com/mbuyakov/SUI.CORE/commit/62f2355))
+
+
+
+
+
+# [5.11.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.9.8...v5.11.0) (2019-09-24)
+
+
+### Bug Fixes
+
+* add labelTemplate for 3DBarChart ([20df06a](https://github.com/mbuyakov/SUI.CORE/commit/20df06a))
+
+
+### Features
+
+* **all:** fixes + colorSettings tab ([63d763c](https://github.com/mbuyakov/SUI.CORE/commit/63d763c))
+
+
+
+
+
+## [5.10.1](https://github.com/mbuyakov/SUI.CORE/compare/v5.10.0...v5.10.1) (2019-09-23)
 
 **Note:** Version bump only for package @smsoft/sui-charts
+
+
+
+
+
+
+# [5.10.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.9.4...v5.10.0) (2019-09-23)
+
+
+### Features
+
+* **all:** fixes + colorSettings tab ([63d763c](https://github.com/mbuyakov/SUI.CORE/commit/63d763c))
+