+++ packages/SUI.META/CHANGELOG.md
@@ -3,9 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [5.9.8](https://github.com/mbuyakov/SUI.CORE/compare/v5.9.7...v5.9.8) (2019-09-23)
+## [5.11.2](https://github.com/mbuyakov/SUI.CORE/compare/v5.11.1...v5.11.2) (2019-09-24)
 
 **Note:** Version bump only for package @smsoft/sui-meta
+
+
+
+
+
+## [5.11.1](https://github.com/mbuyakov/SUI.CORE/compare/v5.11.0...v5.11.1) (2019-09-24)
+
+
+### Bug Fixes
+
+* **asd:** asdsdadssdaasd ([f2204a9](https://github.com/mbuyakov/SUI.CORE/commit/f2204a9))
+* lint ([aa1eddd](https://github.com/mbuyakov/SUI.CORE/commit/aa1eddd))
+
+
+
+
+
+
+# [5.11.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.9.8...v5.11.0) (2019-09-24)
+
+
+### Bug Fixes
+
+* **parameterrowcomponent:** build fix ([2b1c0eb](https://github.com/mbuyakov/SUI.CORE/commit/2b1c0eb))
+
+
+### Features
+
+* **all:** fixes + colorSettings tab ([63d763c](https://github.com/mbuyakov/SUI.CORE/commit/63d763c))
+
+
+
+
+
+## [5.10.3](https://github.com/mbuyakov/SUI.CORE/compare/v5.10.2...v5.10.3) (2019-09-23)
+
+
+### Bug Fixes
+
+* **parameterrowcomponent:** build fix ([2b1c0eb](https://github.com/mbuyakov/SUI.CORE/commit/2b1c0eb))
+
+
+
+
+
+## [5.10.2](https://github.com/mbuyakov/SUI.CORE/compare/v5.10.1...v5.10.2) (2019-09-23)
+
+**Note:** Version bump only for package @smsoft/sui-meta
+
+
+
+
+
+## [5.10.1](https://github.com/mbuyakov/SUI.CORE/compare/v5.10.0...v5.10.1) (2019-09-23)
+
+**Note:** Version bump only for package @smsoft/sui-meta
+
+
+
+
+
+
+# [5.10.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.9.4...v5.10.0) (2019-09-23)
+
+
+### Bug Fixes
+
+* styles ([4e79695](https://github.com/mbuyakov/SUI.CORE/commit/4e79695))
+
+
+### Features
+
+* **all:** fixes + colorSettings tab ([63d763c](https://github.com/mbuyakov/SUI.CORE/commit/63d763c))
+